pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "bathyal-ai"
version = "0.1.0"
description = "Megalodon detection with a production-oriented BioCLIP 2 species classifier."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "dill>=0.4.1",
    "fathomnet",
    "huggingface_hub",
    "numpy>=1.26",
    "open_clip_torch>=2.32.0",
    "pillow>=10.4.0",
    "torch>=2.6.0",
    "ultralytics>=8.3.0",
]

[project.scripts]
bathyal-ai = "bathyal_ai.cli:main"
bathyal-ai-eval = "bathyal_ai.eval_cli:main"
bathyal-ai-train = "bathyal_ai.train_cli:main"
bathyal-ai-ingest = "bathyal_ai.ingest_cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]